requirements.txt
python-telegram-bot==13.7
tweepy==4.14.0
python-dotenv==1.0.0
google-generativeai==0.3.1
Pillow==10.1.0
requests==2.31.0
pytest==7.4.3
flask
openai
asgiref
bs4